requirements.txt
fastapi
uvicorn
pydantic
requests-async
http3
python-dateutil
case-converter